Split sp destinations & path to different files
