+++ src/function/gds/asp_destinations.cpp
@@ -1,15 +1,6 @@
-#include <utility>
-
-#include "binder/expression/node_expression.h"
-#include "common/types/types.h"
-#include "function/gds/auxiliary_state/path_auxiliary_state.h"
-#include "function/gds/bfs_graph.h"
-#include "function/gds/gds_frontier.h"
 #include "function/gds/gds_function_collection.h"
 #include "function/gds/rec_joins.h"
-#include "function/gds_function.h"
-#include "graph/graph.h"
-#include "main/client_context.h"
+#include "binder/expression/node_expression.h"
 #include "processor/execution_context.h"
 
 using namespace kuzu::processor;
@@ -21,9 +12,9 @@ using namespace kuzu::graph;
 namespace kuzu {
 namespace function {
 
-class PathMultiplicities {
+class Multiplicities {
 public:
-    PathMultiplicities(const std::unordered_map<table_id_t, uint64_t>& numNodesMap,
+    Multiplicities(const std::unordered_map<table_id_t, uint64_t>& numNodesMap,
         storage::MemoryManager* mm) {
         for (auto& [tableID, numNodes] : numNodesMap) {
             multiplicityArray.allocate(tableID, numNodes, mm);
@@ -65,9 +56,9 @@ private:
     std::atomic<uint64_t>* curBoundMultiplicities = nullptr;
 };
 
-class AllSPDestinationsAuxiliaryState : public GDSAuxiliaryState {
+class ASPDestinationsAuxiliaryState : public GDSAuxiliaryState {
 public:
-    explicit AllSPDestinationsAuxiliaryState(std::shared_ptr<PathMultiplicities> multiplicities)
+    explicit ASPDestinationsAuxiliaryState(std::shared_ptr<Multiplicities> multiplicities)
         : multiplicities{std::move(multiplicities)} {}
 
     void initSource(common::nodeID_t source) override {
@@ -82,14 +73,14 @@ public:
     }
 
 private:
-    std::shared_ptr<PathMultiplicities> multiplicities;
+    std::shared_ptr<Multiplicities> multiplicities;
 };
 
-class AllSPDestinationsOutputWriter : public RJOutputWriter {
+class ASPDestinationsOutputWriter : public RJOutputWriter {
 public:
-    AllSPDestinationsOutputWriter(main::ClientContext* context, NodeOffsetMaskMap* outputNodeMask,
+    ASPDestinationsOutputWriter(main::ClientContext* context, NodeOffsetMaskMap* outputNodeMask,
         nodeID_t sourceNodeID, std::shared_ptr<PathLengths> pathLengths,
-        std::shared_ptr<PathMultiplicities> multiplicities)
+        std::shared_ptr<Multiplicities> multiplicities)
         : RJOutputWriter{context, outputNodeMask, sourceNodeID},
           pathLengths{std::move(pathLengths)}, multiplicities{std::move(multiplicities)} {
         lengthVector = createVector(LogicalType::UINT16(), context->getMemoryManager());
@@ -114,7 +105,7 @@ public:
     }
 
     std::unique_ptr<RJOutputWriter> copy() override {
-        return std::make_unique<AllSPDestinationsOutputWriter>(context, outputNodeMask,
+        return std::make_unique<ASPDestinationsOutputWriter>(context, outputNodeMask,
             sourceNodeID, pathLengths, multiplicities);
     }
 
@@ -127,13 +118,13 @@ private:
 private:
     std::unique_ptr<ValueVector> lengthVector;
     std::shared_ptr<PathLengths> pathLengths;
-    std::shared_ptr<PathMultiplicities> multiplicities;
+    std::shared_ptr<Multiplicities> multiplicities;
 };
 
-class AllSPDestinationsEdgeCompute : public SPEdgeCompute {
+class ASPDestinationsEdgeCompute : public SPEdgeCompute {
 public:
-    AllSPDestinationsEdgeCompute(SinglePathLengthsFrontierPair* frontierPair,
-        std::shared_ptr<PathMultiplicities> multiplicities)
+    ASPDestinationsEdgeCompute(SinglePathLengthsFrontierPair* frontierPair,
+        std::shared_ptr<Multiplicities> multiplicities)
         : SPEdgeCompute{frontierPair}, multiplicities{std::move(multiplicities)} {};
 
     std::vector<nodeID_t> edgeCompute(nodeID_t boundNodeID, NbrScanState::Chunk& resultChunk,
@@ -162,53 +153,11 @@ public:
     }
 
     std::unique_ptr<EdgeCompute> copy() override {
-        return std::make_unique<AllSPDestinationsEdgeCompute>(frontierPair, multiplicities);
+        return std::make_unique<ASPDestinationsEdgeCompute>(frontierPair, multiplicities);
     }
 
 private:
-    std::shared_ptr<PathMultiplicities> multiplicities;
-};
-
-class AllSPPathsEdgeCompute : public SPEdgeCompute {
-public:
-    AllSPPathsEdgeCompute(SinglePathLengthsFrontierPair* frontiersPair, BFSGraph* bfsGraph)
-        : SPEdgeCompute{frontiersPair}, bfsGraph{bfsGraph} {
-        block = bfsGraph->addNewBlock();
-    }
-
-    std::vector<nodeID_t> edgeCompute(nodeID_t boundNodeID, NbrScanState::Chunk& resultChunk,
-        bool fwdEdge) override {
-        std::vector<nodeID_t> activeNodes;
-        resultChunk.forEach([&](auto nbrNodeID, auto edgeID) {
-            auto nbrLen =
-                frontierPair->getPathLengths()->getMaskValueFromNextFrontier(nbrNodeID.offset);
-            // We should update in 2 cases: 1) if nbrID is being visited
-            // for the first time, i.e., when its value in the pathLengths frontier is
-            // PathLengths::UNVISITED. Or 2) if nbrID has already been visited but in this
-            // iteration, so it's value is curIter + 1.
-            auto shouldUpdate =
-                nbrLen == PathLengths::UNVISITED || nbrLen == frontierPair->getCurrentIter();
-            if (shouldUpdate) {
-                if (!block->hasSpace()) {
-                    block = bfsGraph->addNewBlock();
-                }
-                bfsGraph->addParent(frontierPair->getCurrentIter(), boundNodeID, edgeID, nbrNodeID,
-                    fwdEdge, block);
-            }
-            if (nbrLen == PathLengths::UNVISITED) {
-                activeNodes.push_back(nbrNodeID);
-            }
-        });
-        return activeNodes;
-    }
-
-    std::unique_ptr<EdgeCompute> copy() override {
-        return std::make_unique<AllSPPathsEdgeCompute>(frontierPair, bfsGraph);
-    }
-
-private:
-    BFSGraph* bfsGraph;
-    ObjectBlock<ParentList>* block = nullptr;
+    std::shared_ptr<Multiplicities> multiplicities;
 };
 
 /**
@@ -239,56 +188,13 @@ private:
         auto frontier = getPathLengthsFrontier(context, PathLengths::UNVISITED);
         auto numNodesMap = sharedState->graph->getNumNodesMap(clientContext->getTransaction());
         auto mm = clientContext->getMemoryManager();
-        auto multiplicities = std::make_shared<PathMultiplicities>(numNodesMap, mm);
-        auto outputWriter = std::make_unique<AllSPDestinationsOutputWriter>(clientContext,
+        auto multiplicities = std::make_shared<Multiplicities>(numNodesMap, mm);
+        auto outputWriter = std::make_unique<ASPDestinationsOutputWriter>(clientContext,
             sharedState->getOutputNodeMaskMap(), sourceNodeID, frontier, multiplicities);
         auto frontierPair = std::make_unique<SinglePathLengthsFrontierPair>(frontier);
         auto edgeCompute =
-            std::make_unique<AllSPDestinationsEdgeCompute>(frontierPair.get(), multiplicities);
-        auto auxiliaryState = std::make_unique<AllSPDestinationsAuxiliaryState>(multiplicities);
-        auto gdsState = std::make_unique<GDSComputeState>(std::move(frontierPair),
-            std::move(edgeCompute), std::move(auxiliaryState), sharedState->getOutputNodeMaskMap());
-        return RJCompState(std::move(gdsState), std::move(outputWriter));
-    }
-};
-
-class AllSPPathsAlgorithm final : public RJAlgorithm {
-public:
-    AllSPPathsAlgorithm() = default;
-    AllSPPathsAlgorithm(const AllSPPathsAlgorithm& other) : RJAlgorithm{other} {}
-
-    expression_vector getResultColumns(const function::GDSBindInput& /*bindInput*/) const override {
-        auto rjBindData = bindData->ptrCast<RJBindData>();
-        expression_vector columns;
-        columns.push_back(bindData->getNodeInput()->constCast<NodeExpression>().getInternalID());
-        columns.push_back(bindData->getNodeOutput()->constCast<NodeExpression>().getInternalID());
-        columns.push_back(rjBindData->lengthExpr);
-        if (rjBindData->extendDirection == ExtendDirection::BOTH) {
-            columns.push_back(rjBindData->directionExpr);
-        }
-        columns.push_back(rjBindData->pathNodeIDsExpr);
-        columns.push_back(rjBindData->pathEdgeIDsExpr);
-        return columns;
-    }
-
-    std::unique_ptr<GDSAlgorithm> copy() const override {
-        return std::make_unique<AllSPPathsAlgorithm>(*this);
-    }
-
-private:
-    RJCompState getRJCompState(ExecutionContext* context, nodeID_t sourceNodeID) override {
-        auto clientContext = context->clientContext;
-        auto frontier = getPathLengthsFrontier(context, PathLengths::UNVISITED);
-        auto bfsGraph = getBFSGraph(context);
-        auto rjBindData = bindData->ptrCast<RJBindData>();
-        auto writerInfo = rjBindData->getPathWriterInfo();
-        writerInfo.pathNodeMask = sharedState->getPathNodeMaskMap();
-        auto outputWriter = std::make_unique<SPPathsOutputWriter>(clientContext,
-            sharedState->getOutputNodeMaskMap(), sourceNodeID, writerInfo, *bfsGraph);
-        auto frontierPair = std::make_unique<SinglePathLengthsFrontierPair>(frontier);
-        auto edgeCompute =
-            std::make_unique<AllSPPathsEdgeCompute>(frontierPair.get(), bfsGraph.get());
-        auto auxiliaryState = std::make_unique<PathAuxiliaryState>(std::move(bfsGraph));
+            std::make_unique<ASPDestinationsEdgeCompute>(frontierPair.get(), multiplicities);
+        auto auxiliaryState = std::make_unique<ASPDestinationsAuxiliaryState>(multiplicities);
         auto gdsState = std::make_unique<GDSComputeState>(std::move(frontierPair),
             std::move(edgeCompute), std::move(auxiliaryState), sharedState->getOutputNodeMaskMap());
         return RJCompState(std::move(gdsState), std::move(outputWriter));
@@ -303,18 +209,6 @@ function_set AllSPDestinationsFunction::getFunctionSet() {
 
 GDSFunction AllSPDestinationsFunction::getFunction() {
     auto algo = std::make_unique<AllSPDestinationsAlgorithm>();
-    auto params = algo->getParameterTypeIDs();
-    return GDSFunction(name, std::move(params), std::move(algo));
-}
-
-function_set AllSPPathsFunction::getFunctionSet() {
-    function_set result;
-    result.push_back(std::make_unique<GDSFunction>(getFunction()));
-    return result;
-}
-
-GDSFunction AllSPPathsFunction::getFunction() {
-    auto algo = std::make_unique<AllSPPathsAlgorithm>();
     auto params = algo->getParameterTypeIDs();
     return GDSFunction(name, std::move(params), std::move(algo));
 }
